+++ knotx-core/src/main/java/io/knotx/assembler/FragmentAssemblerHandler.java
@@ -15,7 +15,7 @@
  */
 package io.knotx.assembler;
 
-import io.knotx.fragment.ClientResponse;
+import io.knotx.server.api.ClientResponse;
 import io.knotx.server.api.FragmentsContext;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.vertx.core.Handler;
