+++ knotx-core/src/main/java/io/knotx/fallback/FragmentFallbackHandler.java
@@ -15,9 +15,9 @@
  */
 package io.knotx.fallback;
 
-import io.knotx.fragment.ClientResponse;
 import io.knotx.fragment.Fragment;
 import io.knotx.options.FallbackMetadata;
+import io.knotx.server.api.ClientResponse;
 import io.knotx.server.api.FragmentsContext;
 import io.knotx.snippet.SnippetFragmentsContext;
 import io.netty.handler.codec.http.HttpResponseStatus;
