+++ knotx-core/src/main/java/io/knotx/server/api/FragmentsContext.java
@@ -16,8 +16,6 @@
 package io.knotx.server.api;
 
 import com.google.common.collect.Lists;
-import io.knotx.fragment.ClientRequest;
-import io.knotx.fragment.ClientResponse;
 import io.knotx.fragment.Fragment;
 import io.knotx.fragment.HandlerLogEntry;
 import io.vertx.codegen.annotations.DataObject;
