+++ knotx-core/src/main/java/io/knotx/server/handler/repository/RepositoryRoutingHandlerFactory.java
@@ -15,8 +15,8 @@
  */
 package io.knotx.server.handler.repository;
 
-import io.knotx.fragment.ClientResponse;
 import io.knotx.reactivex.proxy.RepositoryConnectorProxy;
+import io.knotx.server.api.ClientResponse;
 import io.knotx.server.api.FragmentsContext;
 import io.knotx.server.handler.api.RoutingHandlerFactory;
 import io.netty.handler.codec.http.HttpResponseStatus;
