+++ knotx-core/src/main/java/io/knotx/snippet/SnippetFragmentsContext.java
@@ -18,9 +18,9 @@ package io.knotx.snippet;
 import com.google.common.base.Objects;
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
-import io.knotx.fragment.ClientRequest;
-import io.knotx.fragment.ClientResponse;
 import io.knotx.fragment.Fragment;
+import io.knotx.server.api.ClientRequest;
+import io.knotx.server.api.ClientResponse;
 import io.knotx.server.api.FragmentsContext;
 import io.reactivex.Single;
 import io.vertx.codegen.annotations.DataObject;
