+++ server-http/api/src/main/java/io/knotx/server/api/DataObjectsUtil.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.fragment;
+package io.knotx.server.api;
 
 import com.google.common.base.Joiner;
 import com.google.common.base.Objects;
